Convert polynomials to Horner form
This form can be evaluated with just 3 mul and 3 add. Neat!
Moreover, the coefficients W X Y can be precalc'ed and stored instead of
the B C D from the original poly (we already know A/Z because it is the
ending point of the previous Bézier). So this comes at no cost.